Add tf_copts to some C API targets
This is important for the TF_COMPILE_LIBRARY macro definition on Windows, used for marking functions with the dllexport attribute instead of dllimport. Not having it results in compiler warnings and seemingly "undefined reference to symbol" errors when not linking with another target that has the symbol generated with dllexport.

PiperOrigin-RevId: 489563722