Allow dumping MLIR files during preprocess passes for TF quantization.
Currently MLIR dumps are only enabled for the main quantization passes and are omitted for the preprocessing passes where variable freezing happens.
To enhance debuggability for the preprocessing passes as well, this change allows producing MLIR dump files for the preprocess passes.

Functionalities for MLIR dump are moved to a separate file `mlir_dump.cc` for modularity.

PiperOrigin-RevId: 489598822