Add new and delete operators to the stable delegate linker script.
This adds new and delete operators to the list of symbols that we keep
when building the sample stable delegate.  Those end up defined as weak
symbols in the stable delegate, and we need to ensure those get bound to
the same symbol for the stable delegate as for the main executable,
since data that is allocated by the stable delegate may be deallocated
elsewhere.

PiperOrigin-RevId: 490873417