Add separate tracking for HBM BW utilization.
Previously, we were combining the sum of all memory spaces and dividing that sum by the sum of peak BWs.
Now, we report both the old utilization number and the HBM utilization number separately.

In the future, we probably want to remove the aggregated utilization number and report the on-chip utilization numbers (rd/wr) in addition to the hbm utilization.

Note, this CL only modifies the backend, a child CL will modify the frontend as well to surface the new information to the UI.

PiperOrigin-RevId: 491984750