[GmlSt] Fold gml_st.materialize that has source and result of the same shape.
Currently we only fold MaterializeOp inside of ForOp, with the source of the former being an output argument of the latter.

The folding is needed by transfer_read/write hoisting, which will also be in vectorization pass. Hence it makes more sense to put the pattern in vectorization as well.

Note: the util functions definition/declaration are also moved into transforms.cc/.h.
PiperOrigin-RevId: 492002580