[XLA:GPU] Simplify EmitF32ToBF16 in GpuElementalIrEmitter for Cuda capability >= 8.0
We are using the more complex default operation for Cuda capability < 8.0,
and a faster hardware solution for Cuda capability >= 8.0.

We are now running the elemental_ir_emitter_test on gpu_sm70only and gpu_sm80only in addition to cpu and gpu.

We also pass down the IrEmitterContext to GpuElementalIrEmitter to be able to check the Cuda capability.

PiperOrigin-RevId: 493008856