Disable Scorecards workflow temporarily.
The workflow has 2 issues: one that is making it fail and a more recent one resulting in auth failure. While we know the trigger for the credential one and GOSST is working on getting a fix, we don't yet have a solution for the first issue. So, disabling this workflow and only allowing manual triggers to test for fixes.

PiperOrigin-RevId: 493954067