Add a pass that creates a RestoreV2Op for every variable initialization pattern in the initializer function.
Adds a new pass called `CreateRestoreOpPass` that inserts a new `RestoreV2Op` that reads the tensors that initialize variables from a checkpoint and uses the loaded tensors to initialize the variables.
After this pass, the `tf.Const` ops that are originally used to initialize the variables are replaced by the results of `tf.RestoreV2Op`.

PiperOrigin-RevId: 494647727