Merge Flex operator profiling into main profile
A flag kTfLiteDelegateFlagsPerOperatorProfiling is set when Flex delegate is created, disabling the profiling of a macro operator TfLiteFlexDelegate that grouped all the delegate operators.
Then, when delegate operator's profiler is initialized, the EventType is set to Profiler::EventType::DELEGATE_PROFILED_OPERATOR_INVOKE_EVENT, making the statistics of each delegate operators appear in the Operator-wise Profiling section and not in the Delegate internal section.

PiperOrigin-RevId: 494964831