Remove unused passes in mlir::odml
This is part 2 of 5 of the CL chain that migrates odml_to_stablehlo from producing MHLO to producing StableHLO.

To further prepare for the migration, I figured I'd trim some of the unused code to make the migration easier. All these passes looked unused and superseded by the current plan of record.

PiperOrigin-RevId: 495151891