Inline PopulateTFToMhloPatterns into tf_stablehlo_pass.h
This is part 4 of 5 of the CL chain that migrates odml_to_stablehlo from producing MHLO to producing StableHLO.

Now that we've cleaned up some unused passes in part 2, PopulateTFToMhloPatterns has only one use, so I've inlined it into the specific pass where it's used. This will simplify the migration.

PiperOrigin-RevId: 495234512