Merge restore op initializer function to the main graph for TF quantizer.
In order to export the restore op (`RestoreV2`) created by the `CreateRestoreOpPass`, also merge the initializer function with initializer_type == "restore_op" within the `MergeInitializerFunctionOpsToMainPass`.
The output control value (`tf_executor::ControlType`) will correspond to the `tf.NoOp` in the main Graph that has control dependencies on the variable restoration.

PiperOrigin-RevId: 495734093