Disable TF session static graph optimizations in compiler/xrt/tests.
This change explicitly sets the value of `disable_optimize_for_static_graph` to
true in TF sessions to disable optimizations made to the static graph. This
will allow the users of this particular session to continue making use of its
`Extend()` API which is incompatible with the optimization feature.

Note that this CL is effectively a no-op since today the optimization feature
is disabled by default.

This will help support the calls to Extend that the
tensorflow::XrtClientSession makes internally in its tensorflow::ClientSession.
PiperOrigin-RevId: 496707037